refactor: move sign out of origo
